move blockIndex functions to shared blockIndex.ts file
