+++ packages/era/src/index.ts
@@ -7,6 +7,7 @@ export * from './e2hs/index.ts'
 export * from './exportHistory.ts'
 export * from './snappy.ts'
 export * from './types.ts'
+export * from './blockIndex.ts'
 
 export function readBinaryFile(path: string) {
   return new Uint8Array(readFileSync(path))
